+++ crates/resvg/examples/custom_usvg_tree.rs
@@ -53,7 +53,7 @@ fn main() {
 
     let rtree = resvg::Tree::from_usvg(&tree);
 
-    let pixmap_size = tree.size.to_screen_size();
+    let pixmap_size = resvg::ScreenSize::from_usvg(rtree.size);
     let mut pixmap = tiny_skia::Pixmap::new(pixmap_size.width(), pixmap_size.height()).unwrap();
     rtree.render(tiny_skia::Transform::default(), &mut pixmap.as_mut());
     pixmap.save_png("out.png").unwrap();